requirements.txt
opencv-python
tensorflow-gpu==1.3.0